+++ android/src/main/java/com/horcrux/svg/EllipseShadowNode.java
@@ -53,11 +53,11 @@ class EllipseShadowNode extends RenderableShadowNode {
     @Override
     protected Path getPath(Canvas canvas, Paint paint) {
         Path path = new Path();
-        float cx = relativeOnWidth(mCx);
-        float cy = relativeOnHeight(mCy);
-        float rx = relativeOnWidth(mRx);
-        float ry = relativeOnHeight(mRy);
-        RectF oval = new RectF(cx - rx, cy - ry, cx + rx, cy + ry);
+        double cx = relativeOnWidth(mCx);
+        double cy = relativeOnHeight(mCy);
+        double rx = relativeOnWidth(mRx);
+        double ry = relativeOnHeight(mRy);
+        RectF oval = new RectF((float) (cx - rx), (float) (cy - ry), (float) (cx + rx), (float) (cy + ry));
         path.addOval(oval, Path.Direction.CW);
 
         return path;
